requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
python-dotenv
pydantic-settings
alembic
python-jose
passlib
loguru
geopy